feat: WSP design system polish, PricingTab import-from-WBS, and demo fixes
- PricingTab: complete WSP design system restyle (wsp-card, wsp-table,
  wsp-btn-primary), add "↑ From WBS" import button with phaseHours()
  distribution logic, CAD currency formatter, grand total footer
- OverviewTab, DeliverablesTab, DrawingsTab: minor styling/bug fixes
  surfaced during live demo walkthrough
- PeopleTab: polling interval cleanup and CV apply flow fixes
- Backend: cv_fetcher agent improvements, wbs schema total_cost field,
  auth schema cleanup, docker-compose port/volume fixes

+++ backend/app/schemas/wbs.py
@@ -28,9 +28,9 @@ class WBSItemOut(BaseModel):
     wbs_code: str
     description: Optional[str]
     phase: Optional[str]
-    hours: Decimal
-    unit_rate: Decimal
-    total_cost: Decimal
+    hours: float
+    unit_rate: float
+    total_cost: float
     order_index: int = 0
 
     model_config = {"from_attributes": True}
